Update user IBAN
