+++ cyclos_configuration/scripts/emandates/Operation_InternalCreate_Callback.groovy
@@ -1,6 +1,5 @@
 import org.cyclos.entities.system.CustomFieldPossibleValue
 import org.cyclos.entities.system.ExternalRedirectExecution
-import org.cyclos.entities.users.User
 import org.cyclos.impl.system.EntityBackedParameterStorage
 import org.cyclos.model.ValidationException
 import org.cyclos.model.utils.RequestInfo
@@ -11,27 +10,13 @@ ExternalRedirectExecution execution = binding.execution
 ObjectMapper objectMapper = binding.objectMapper
 RequestInfo request = binding.request
 def transactionId = request.getParameter('transactionId')
-
 def eMandates = new EMandates(binding)
 def storage = new EntityBackedParameterStorage(objectMapper, execution)
-def fields = eMandates.callback(storage, transactionId)
 
-String status = (fields.status as CustomFieldPossibleValue).internalName
-if ('success' == status && fields.owner instanceof User) {
-    // Store the iban from the eMandate in the user profile.
-    try{
-	    def usrDTO = userService.load((fields.owner as User).id)
-        def usr = scriptHelper.wrap(usrDTO)
-        usr.iban = fields.iban
-        userService.save(usrDTO)
-    } catch (ValidationException vE) {
-        return scriptParameters["errorSaveIBAN"] + " '${vE.validation?.firstError}'."
-    }
-// If the iban in the eMandate record is different than the iban we have for this user, inform our financial admin by mail.
-// Implement the new Utils library before using the next lines.
-//	if ( ! utils.isIbansEqual(usr.iban, fields.iban) ) {
-//		sendMailToAdmin("Incassomachtiging van afwijkend iban", utils.prepareMessage("eMandateDifferentBankAccount", ["user": usr.name]))
-//	}
+try{
+    def fields = eMandates.callback(storage, transactionId)
+    String status = (fields.status as CustomFieldPossibleValue).internalName
+    return scriptParameters["result.${status}"]
+}catch (ValidationException vE) {
+    return vE.getMessage()
 }
-
-return scriptParameters["result.${status}"]
